# frontend/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TrainingCurvesResponse": {
        "properties": {
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/TrainingCheckpointInfo"
            },
            "title": "Checkpoints",
            "type": "array"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "test_points": {
            "items": {
              "$ref": "#/components/schemas/TrainingTestPoint"
            },
            "title": "Test Points",
            "type": "array"
          },
          "training_epoch_points": {
            "items": {
              "$ref": "#/components/schemas/TrainingPoint"
            },
            "title": "Training Epoch Points",
            "type": "array"
          },
          "training_points": {
            "items": {
              "$ref": "#/components/schemas/TrainingPoint"
            },
            "title": "Training Points",
            "type": "array"
          }
        },
        "required": [
          "job_id",
          "training_points",
          "training_epoch_points",
          "test_points",
          "checkpoints"
        ],
        "title": "TrainingCurvesResponse",
        "type": "object"
      },
...
      "TrainingJobCreateRequest": {
        "properties": {
          "auto_stop_enabled": {
            "default": false,
            "title": "Auto Stop Enabled",
            "type": "boolean"
          },
          "auto_stop_metric": {
            "default": "f1",
            "enum": [
              "accuracy",
              "precision",
              "recall",
              "f1",
              "pr_auc"
            ],
            "title": "Auto Stop Metric",
            "type": "string"
          },
          "auto_stop_min_epochs": {
            "default": 1,
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Auto Stop Min Epochs",
            "type": "integer"
          },
          "auto_stop_threshold": {
            "default": 0.98,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Auto Stop Threshold",
            "type": "number"
          },
          "batch_size": {
            "default": 256,
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Batch Size",
            "type": "integer"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "epochs": {
            "default": 0,
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Epochs",
            "type": "integer"
          },
          "eval_interval": {
            "default": 1,
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Eval Interval",
            "type": "integer"
          },
          "gpu_id": {
            "minimum": 0.0,
            "title": "Gpu Id",
            "type": "integer"
          },
          "input_representation": {
            "const": "embedding",
            "default": "embedding",
            "title": "Input Representation",
            "type": "string"
          },
          "keep_last_epochs": {
            "default": 5,
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Keep Last Epochs",
            "type": "integer"
          },
          "learning_rate": {
            "default": 0.0002,
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Learning Rate",
            "type": "number"
          },
          "max_test_samples": {
            "anyOf": [
              {
                "maximum": 100000000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Test Samples"
          },
          "max_train_samples": {
            "anyOf": [
              {
                "maximum": 100000000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Train Samples"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "num_workers": {
            "default": 8,
            "maximum": 128.0,
            "minimum": 0.0,
            "title": "Num Workers",
            "type": "integer"
          },
          "prepare_workers": {
            "anyOf": [
              {
                "maximum": 128.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prepare Workers"
          },
          "resume_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume From"
          },
          "scenarios": {
            "items": {
              "type": "string"
            },
            "title": "Scenarios",
            "type": "array"
          },
          "seed": {
            "default": 42,
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "simulator": {
            "default": "modflow",
            "title": "Simulator",
            "type": "string"
          },
          "test_batch_size": {
            "default": 256,
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Test Batch Size",
            "type": "integer"
          },
          "test_ratio": {
            "default": 0.1,
            "maximum": 0.9,
            "minimum": 0.0,
            "title": "Test Ratio",
            "type": "number"
          },
          "weight_decay": {
            "default": 0.01,
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Weight Decay",
            "type": "number"
          }
        },
        "required": [
          "gpu_id"
        ],
        "title": "TrainingJobCreateRequest",
        "type": "object"
      },
...
      "TrainingJobSummary": {
        "properties": {
          "artifact_root": {
            "title": "Artifact Root",
            "type": "string"
          },
          "avg_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Loss"
          },
          "config": {
            "$ref": "#/components/schemas/TrainingJobConfig"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "eta_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Seconds"
          },
          "exit_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Reason"
          },
          "global_step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Step"
          },
          "gpu_id": {
            "title": "Gpu Id",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "latest_epoch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Epoch"
          },
          "latest_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainingMetricsSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Step"
          },
          "latest_test_epoch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Test Epoch"
          },
          "log_path": {
            "title": "Log Path",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "pipeline_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Stage"
          },
          "router_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainingMetricsSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "router_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Router Status"
          },
          "run_dir": {
            "title": "Run Dir",
            "type": "string"
          },
          "scenarios": {
            "items": {
              "type": "string"
            },
            "title": "Scenarios",
            "type": "array"
          },
          "simulator": {
            "title": "Simulator",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "starting",
              "running",
              "evaluating",
              "stopping",
              "done",
              "error",
              "terminated",
              "external_terminated"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps_per_epoch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps Per Epoch"
          },
          "steps_per_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps Per Sec"
          },
          "stop_requested": {
            "default": false,
            "title": "Stop Requested",
            "type": "boolean"
          },
          "stop_requested_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Requested At"
          },
          "text2comp_dataset_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Dataset Path"
          },
          "text2comp_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Error Message"
          },
          "text2comp_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Job Id"
          },
          "text2comp_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainingText2CompMetricsSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "text2comp_model_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Model Path"
          },
          "text2comp_output_dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Output Dim"
          },
          "text2comp_quality_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Quality Passed"
          },
          "text2comp_run_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Run Dir"
          },
          "text2comp_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Status"
          },
          "text2comp_target_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text2Comp Target Source"
          },
          "uploaded_expert_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded Expert Id"
          },
          "uploaded_expert_input_dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded Expert Input Dim"
          },
          "uploaded_expert_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded Expert Name"
          }
        },
        "required": [
          "job_id",
          "name",
          "status",
          "simulator",
          "scenarios",
          "gpu_id",
          "created_at",
          "artifact_root",
          "run_dir",
          "log_path",
          "config"
        ],
        "title": "TrainingJobSummary",
        "type": "object"
      },
      "TrainingLogResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "lines": {
            "items": {
              "type": "string"
            },
            "title": "Lines",
            "type": "array"
          }
        },
        "required": [
          "job_id",
          "lines"
        ],
        "title": "TrainingLogResponse",
        "type": "object"
      },
      "TrainingMetricsSummary": {
        "properties": {
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "f1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "F1"
          },
          "pr_auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Auc"
          },
          "precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precision"
          },
          "recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall"
          }
        },
        "title": "TrainingMetricsSummary",
        "type": "object"
      },
...
      "TrainingQuickJobCreateRequest": {
        "properties": {
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "gpu_id": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "resume_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume From"
          },
          "scenarios": {
            "items": {
              "type": "string"
            },
            "title": "Scenarios",
            "type": "array"
          },
          "seed": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "simulator": {
            "default": "modflow",
            "title": "Simulator",
            "type": "string"
          },
          "uploaded_expert_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded Expert Id"
          }
        },
        "title": "TrainingQuickJobCreateRequest",
        "type": "object"
      },
...
      "TrainingText2CompMetricsSummary": {
        "properties": {
          "mae": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae"
          },
          "mse": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mse"
          },
          "normalized_rmse": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Rmse"
          },
          "r2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "R2"
          }
        },
        "title": "TrainingText2CompMetricsSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Piern Unified API",
    "version": "3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/training/jobs": {
      "get": {
        "operationId": "get_training_jobs_api_training_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrainingJobSummary"
                  },
                  "title": "Response Get Training Jobs Api Training Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Training Jobs",
        "tags": [
          "training"
        ]
      },
      "post": {
        "operationId": "create_training_job_api_training_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingJobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Training Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_training_job_api_training_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Training Job",
        "tags": [
          "training"
        ]
      },
      "get": {
        "operationId": "get_training_job_api_training_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Training Job",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/curves": {
      "get": {
        "operationId": "get_training_curves_api_training_jobs__job_id__curves_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_points",
            "required": false,
            "schema": {
              "default": 2000,
              "maximum": 10000,
              "minimum": 100,
              "title": "Max Points",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingCurvesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Training Curves",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/logs": {
      "get": {
        "operationId": "get_training_logs_api_training_jobs__job_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 2000,
              "minimum": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Training Logs",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/jobs/{job_id}/stop": {
      "post": {
        "operationId": "stop_training_job_api_training_jobs__job_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Training Job",
        "tags": [
          "training"
        ]
      }
    },
...
    "/api/training/quick-jobs": {
      "post": {
        "operationId": "create_quick_training_job_api_training_quick_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingQuickJobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Quick Training Job",
        "tags": [
          "training"
        ]
      }
    },
...
  }
}